Sanitizing return types of void/Response methods.
For REST methods with a void return type or methods returning
javax.ws.rs.core.Response (which means that it might be returning
nothing or returning something but also tampering the response headers),
we are now returning retrofit.client.Response per default.

For methods returning javax.ws.rs.core.Response we are also looking for
a possible ApiOperation annotation and check if a response class is
defined in it, using it as the return type for the method if it is.